requirements.txt
blobfile==2.1.1
ipython==8.12.3
lightning==2.1.4
torch-xla==2.2.0
lmdb==1.4.1
lpips==0.1.4
mpi4py==3.1.5
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
piq==0.8.0
pytorch_msssim==1.0.0
PyYAML==6.0.1
Requests==2.32.3
rich==13.7.1
scikit_learn==1.5.1
scipy==1.13.1
setuptools==69.0.3
scikit-image==0.24.0
tensorflow==2.17.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
